Moved Picker Handler tests between different classes
