+++ src/Core/tests/DeviceTests/Handlers/Picker/PickerHandlerTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using Microsoft.Maui.DeviceTests.Stubs;
+﻿using Microsoft.Maui.DeviceTests.Stubs;
 using Microsoft.Maui.Handlers;
 
 namespace Microsoft.Maui.DeviceTests
@@ -10,18 +9,5 @@ namespace Microsoft.Maui.DeviceTests
 		public PickerHandlerTests(HandlerTestFixture fixture) : base(fixture)
 		{
 		}
-
-#if MONOANDROID
-		[Fact(DisplayName = "Title Initializes Correctly")]
-		public async Task TitleInitializesCorrectly()
-		{
-			var picker = new PickerStub
-			{
-				Title = "Select an Item"
-			};
-
-			await ValidatePropertyInitValue(picker, () => picker.Title, GetNativeTitle, picker.Title);
-		}
-#endif
 	}
 }
